fix: fix cart-products click
